+++ custom_components/securityspy/entity.py
@@ -93,6 +93,7 @@ class SecuritySpyEntity(Entity):
 
     async def async_added_to_hass(self):
         """When entity is added to hass."""
+        await super().async_added_to_hass()
         self.async_on_remove(
             self.secspy_data.async_subscribe_device_id(
                 self._device_id, self.async_write_ha_state